refactor: Move the handleMath to folder utils
