+++ src/App.jsx
@@ -1,45 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import GlobalStyles from './GlobalStyles';
 import StyledMain from './App.style';
 import {
   BotContainer, MidContainer, TopContainer, History, Current,
 } from './components';
 import { ButtonValueProvider, useButtonValue } from './context/ButtonValue';
+import handleMath from './utils/handleMath';
 
 function Display() {
-  const { getButtonValue, resetValues } = useButtonValue();
+  const { getButtonValue } = useButtonValue();
   const currentValue = getButtonValue();
-
-  function handleMath(value) {
-    const [memoValue, setMemoValue] = useState({
-      current: '0',
-      history: '',
-    });
-    let isValueValid = false;
-    let internalValue = value;
-
-    if (typeof value === 'number') {
-      if (memoValue.current === '0') {
-        internalValue = value;
-      } else {
-        internalValue = `${memoValue.current}${value}`;
-      }
-      isValueValid = true;
-    }
-
-    useEffect(() => {
-      if (isValueValid) {
-        setMemoValue({
-          current: internalValue.toString(),
-          history: '',
-        });
-        resetValues();
-      }
-      isValueValid = false;
-    }, [isValueValid]);
-
-    return memoValue.current;
-  }
 
   const current = handleMath(currentValue);
